+++ CHANGELOG.md
@@ -1,6 +1,15 @@
 # Changelog
 Used [Keep a Changelog](https://keepachangelog.com/en/1.0.0/) and [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [v1.35.0] - 2023-04-20
+### Changed
+- Method `sbpCheck`
+- Method `sbpClientCheck`
+### Added
+- Method `sbpCheckSync`
+- Method `sbpClientCheckSync`
+- Method `sbpGetCheckStatus`
+
 ## [v1.34.0] - 2022-09-09
 ### Added
 - Method `beneficiaryBalanceCorrection`